+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+**Version 0.0.5**
+
+- bugfix: "source" field for Purchase and PurchaseItem should be optional
+
+**Version 0.0.4**
+
+- "source" field for Purchase and PurchaseItem
+
 **Version 0.0.3**
 
 - New events: Purchase and PurchaseItem